Voice-Controlled-Robot/pyproject.toml
[project]
name = "dobot-vision-control"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[google]~=1.4",
    "livekit-agents[silero,turn-detector]~=1.4",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-agents[images]",
    "python-dotenv",
    "pydobot>=1.3.2",
]
